Parse push event data partially
The era of GitHub Action v1, we can parse _push_ event data json
with `*github.PushEvent` provided by go-github.

However, since GitHub Action v2, we cannot use the same way to parse it
because the json's _timestamp_ field format has been changed to
`"8/21/2019 6:06:04 AM"` and go-github does not expect its format.
(See go-github's issue 1254)

Fortunately, we don't require to parse all of its json.
We only need a part of it.

Thus we use self-defined type to unmarshal json.

For the future, if go-github implements the type to support events json
by GitHub Actions v2, it would be better to move to them.